requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
ccxt==4.4.32
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mysqlclient==2.2.6
numpy==2.1.3
pillow==11.0.0
propcache==0.2.0
pyasn1==0.6.1
pycares==4.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.2